classification/requirements.txt
numpy
torch
torchvision
pytorch_lightning
opencv-python
facenet_pytorch
imgaug
dvc[gdrive]
hydra-core
scikit-learn
torchsummary
